requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
gdown
google-generativeai
transformers
torch